+++ src/product/tool-definitions.js
@@ -64,6 +64,7 @@ import { pdfDocumentToolDefinitions } from './definitions/pdf-document-tools.js'
 import { pdfProtectToolDefinitions } from './definitions/pdf-protect-tool.js';
 import { pdfSignToolDefinitions } from './definitions/pdf-sign-tool.js';
 import { pdfImageExtractorToolDefinitions } from './definitions/pdf-image-extractor-tool.js';
+import { textToPdfToolDefinitions } from './definitions/text-to-pdf-tool.js';
 import { memeToolDefinitions } from './definitions/meme-generator-tool.js';
 import { pdfContentToolDefinitions } from './definitions/pdf-content-tools.js';
 import { pdfEditorToolDefinitions } from './definitions/pdf-editor-tools.js';
@@ -1119,6 +1120,7 @@ const allToolDefinitions = Object.freeze({
     ...pdfProtectToolDefinitions,
     ...pdfSignToolDefinitions,
     ...pdfImageExtractorToolDefinitions,
+    ...textToPdfToolDefinitions,
     ...memeToolDefinitions,
     ...pdfContentToolDefinitions,
     ...pdfEditorToolDefinitions,